requirements.txt
# ── Framework ──────────────────────────────────────────────────────────────────
Django>=4.2,<6.0

# ── Images ─────────────────────────────────────────────────────────────────────
Pillow>=10.0

# ── Filtres ────────────────────────────────────────────────────────────────────
django-filter>=23.0

# ── Variables d'environnement ──────────────────────────────────────────────────
python-decouple>=3.8

# ── Base de données PostgreSQL ─────────────────────────────────────────────────
psycopg2-binary>=2.9

# ── Serveur WSGI production ────────────────────────────────────────────────────
gunicorn>=21.0

# ── Fichiers statiques (sans Nginx) ───────────────────────────────────────────
whitenoise>=6.6
